requirements.txt
numpy
matplotlib
Pillow
scikit-image
scikit-learn
imbalanced-learn
gradio